🛠️ refactor: consolidate Basic.Reject/Basic.Nack into ack.go and remove reject.go
Move delivery-related handlers (Basic.Reject, Basic.Nack) into ack.go to centralize ack/nack/reject logic and delete the now-redundant reject.go. Preserves requeue behavior, persistence cleanup, and existing TODO for dead-lettering.
